pyproject.toml
[tool.poetry]
name = "django-graphql-music-app"
version = "0.1.0"
description = ""
authors = ["Andy Tran <antr900@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.7"
djangorestframework = "^3.11.0"
whitenoise = "^5.1.0"
gunicorn = "^20.0.4"
django-heroku = "^0.3.1"
django-cors-headers = "^3.4.0"
graphene_django = "^2.11.0"
django-graphql-jwt = "^0.3.1"
django_heroku = "^0.3.1"
python-dotenv = "^0.13.0"
dj_database_url = "^0.5.0"
dj-database-url = "^0.5.0"
psycopg2 = "^2.8.5"
psycopg2-binary = "^2.8.5"
django-cursor-pagination = "^0.1.4"
six = "^1.15.0"
django-six = "^1.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
